scraper.py
import requests
from bs4 import BeautifulSoup
import time
import random

def scrape_court_data(search_term):
    """
    Placeholder function to simulate scraping court data.
    IMPORTANT: You MUST replace this with actual web scraping logic
    tailored to the specific court website you are targeting.

    Args:
        search_term (str): The term to search for (e.g., a case number, party name).

    Returns:
        list: A list of dictionaries, where each dictionary represents a case.
              Each case should have 'case_number', 'party_names', 'filing_date', 'status'.
    """
    print(f"Attempting to scrape data for: {search_term}")

    # --- Ethical Scraping Practices ---
    # 1. Respect robots.txt: Always check the target website's robots.txt file.
    #    E.g., https://www.example-court.com/robots.txt
    # 2. Rate Limiting: Do not send too many requests too quickly.
    #    Add delays between requests.
    time.sleep(random.uniform(2, 5)) # Random delay between 2 and 5 seconds

    # --- Replace 'https://www.example-court-data.com/search' with the actual URL ---
    # This is a dummy URL and will not work for real scraping.
    # You will need to inspect the real court website's search page and
    # understand its request parameters and HTML structure.
    base_url = "https://www.example-court-data.com/search"
    headers = {
        'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/100.0.4896.127 Safari/537.36'
    }
    # Parameters for the GET request (adjust based on target website's search form)
    params = {'query': search_term, 'page': 1} # Example parameters

    try:
        response = requests.get(base_url, headers=headers, params=params, timeout=15)
        response.raise_for_status() # Raise an HTTPError for bad responses (4xx or 5xx)
        soup = BeautifulSoup(response.text, 'html.parser')

        scraped_results = []

        # --- DUMMY DATA / PLACEHOLDER PARSING LOGIC ---
        # You need to replace this with actual BeautifulSoup parsing logic
        # based on the HTML structure of the court website you are scraping.
        # Example: Look for specific divs, tables, or list items that contain case info.

        # For demonstration, let's create some dummy data
        if "test_case_123" in search_term:
            scraped_results.append({
                'case_number': '123-ABC-456',
                'party_names': 'John Doe vs. Jane Smith',
                'filing_date': '2023-01-15',
                'status': 'Closed'
            })
        elif "another_case" in search_term:
            scraped_results.append({
                'case_number': '789-XYZ-012',
                'party_names': 'Acme Corp vs. Beta Ltd',
                'filing_date': '2024-03-20',
                'status': 'Active'
            })
        else:
            # Simulate finding some generic results
            for i in range(random.randint(1, 3)): # Simulate 1-3 results
                scraped_results.append({
                    'case_number': f'GEN-{search_term}-{i+1}',
                    'party_names': f'Party A {i} vs. Party B {i}',
                    'filing_date': f'202{random.randint(0,4)}-{random.randint(1,12):02d}-{random.randint(1,28):02d}',
                    'status': random.choice(['Active', 'Closed', 'Pending'])
                })
        # --- END DUMMY DATA / PLACEHOLDER PARSING LOGIC ---

        return scraped_results

    except requests.exceptions.HTTPError as e:
        print(f"HTTP Error during scraping: {e.response.status_code} - {e.response.reason}")
        print(f"Response content: {e.response.text[:200]}...") # Print first 200 chars of response
        return []
    except requests.exceptions.ConnectionError as e:
        print(f"Connection Error during scraping: {e}")
        return []
    except requests.exceptions.Timeout as e:
        print(f"Timeout Error during scraping: {e}")
        return []
    except requests.exceptions.RequestException as e:
        print(f"General Request Error during scraping: {e}")
        return []
    except Exception as e:
        print(f"An unexpected error occurred in scraper.py: {e}")
        return []

if __name__ == '__main__':
    # Example of how to test the scraper directly
    print("--- Testing scraper.py directly ---")
    test_term_1 = "test_case_123"
    data_1 = scrape_court_data(test_term_1)
    print(f"\nScraped data for '{test_term_1}':")
    for item in data_1:
        print(item)

    test_term_2 = "random_query"
    data_2 = scrape_court_data(test_term_2)
    print(f"\nScraped data for '{test_term_2}':")
    for item in data_2:
        print(item)
